feat: define missing user in docker container
Define the missing ${USER} variable in the entrypoint-user.sh script to ensure it is set correctly in the Docker container.

This commit adds a new line of code that sets the ${USER} variable to "${USERNAME}".

The purpose of this change is to properly configure the game server by checking if the specified file exists and setting up necessary variables accordingly.